+++ libs/server/auth/feature/src/lib/auth.module.ts
@@ -1,13 +1,18 @@
 import { Module } from '@nestjs/common';
 
 import { AuthController } from './auth.controller';
-import { RefreshJwtGuard } from './guards';
+import { MfaPendingGuard, RefreshJwtGuard } from './guards';
 import { JwtAuthModule } from './jwt-auth.module';
 import { LocalStrategy, RefreshJwtStrategy } from './strategies';
 
 @Module({
   imports: [JwtAuthModule],
   controllers: [AuthController],
-  providers: [LocalStrategy, RefreshJwtStrategy, RefreshJwtGuard],
+  providers: [
+    LocalStrategy,
+    RefreshJwtStrategy,
+    RefreshJwtGuard,
+    MfaPendingGuard,
+  ],
 })
 export class AuthModule {}